Remove old versions of EnergyEmissionsModule
